requirements.txt
# Framework web
Flask==3.0.0
Werkzeug==3.0.1

# Base de datos
mysql-connector-python==8.2.0

# Requests para APIs
requests==2.31.0

# Variables de entorno
python-dotenv==1.0.0

# Utilidades para finanzas
yfinance==0.2.33

# Formateo de fechas
python-dateutil==2.8.2

# Desarrollo (opcional)
flask-debugtoolbar==0.14.1

# Testing (opcional)
pytest==7.4.3
pytest-flask==1.3.0